Initial commit of working ported Chip8 interpreter for Gamebuino
The emulation works well, but at a framerate that is a little too
slow to be playable for games.  Also, I haven't implemented input
or loading games from SD.

Since the Chip8 has 4k of ram and the Gamebuino has only 2k, I had
to page the Chip8 ram to the SD card.  Unfortunately, since the
SD library is so large, I had to make the buffer very small: 156
bytes.  I chose 156 because it fit into 4096 26 times, so I could
save each page to a unique file on the SD card, indexed by a
letter of the alphabet.  128 would have required too many files,
and I was getting ram corruption at 256.

I chose to page out to separate files to keep the implementation
simple.  I didn't want to have to deal with managing complex I/O.