Completed adding game loading
Games now load successfully from the SD card.  It is very
inefficient because it loads one byte at a time.  I need to load
in chunks but I got lazy about calculating it, so it's slower for
now.

I had some weird corruption issues that indicated that some of
these functions have enough local vars and arguments that I'm
pushing the edges of memory.  I need to be very careful about
RAM usage for any future features, or else the application will
crash in unpredictable ways.