Added button mapping and ability to control chip8 games
Before, there was no way to interact with the emulation, but
now input seems to work.  There was a VERY weird bug in which a
function that was called simply returned the wrong result, but if
the result was saved in an intermediate variable, it worked.  Very
frustrating.

Since the chip8 has 16 buttons and the Gamebuino only has 7, it is
impossible to play all chip8 games.  However, most of them only use
a small subset of the keys.  I made it so that the Gamebuino
buttons can be quickly and easily remapped for different games.  I
reserved the C button for the menu, so there are only 6 buttons
that can be used by the user to interact with chip8 programs.