Removed some bitwise &s, changed random function
I changed the random function from libc rand to Arduino random.
There are some roms that do not run correctly and I am attempting
to determine what instructions are not behaving correctly.  I
thought it might be an issue with random number generation, but
the changes appear to have not changed emulator behavior.