Fixed some bugs from the original interpreter
It appears that the semantics of a few of the instructions were
wrong.  I was testing with some test chip8 roms, and some of them
were not working correctly.  These changes make the test roms
produce the right output, but there are still a few more complex
roms that don't seem to work right.